build: Drop py3.9 support
It has been necessary to reduce deps and their version range, in order
to be reduce the complexity of constraints resolution.

In any case, all the version ranges have been restricted to compatible
ones, and all the deps dropped are apparently unused in the code.

Now, it locks in less than 10s